LPS-107874 Refactor, move enum to its own class
- Since it's going to be used from different packages
